prepare verity-server usable image
Adds necessary docker builds to .gitlab-ci.yml and adds a second docker
image that depends on the first but copies the files directly into the
image.
Updates the node-vcx-wrapper dependency. We eventually need to move to
the version published to npm but there are several scripts that depend
on the url being in the package.json.
Updates the provision_sdk.py script to include hard-coded information
for using with the mock agency.